+++ src/parser.yy
@@ -9,6 +9,8 @@
 #include "lexer.h"
 #include "parser.h"
 #include "ast.h"
+#include "ast/syntax.h"
+#include "ast/util.h"
 
 FILE *yyin = NULL;
